requirements.txt

absl-py==2.3.1
accelerate==1.12.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
args==0.1.0
asttokens==3.0.1
async-timeout==5.0.1
attrs==25.4.0
braceexpand==0.1.7
brotli==1.2.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
clint==0.5.1
contourpy==1.3.2
cycler==0.12.1
datasets==4.4.1
decorator==5.2.1
diffusers==0.35.2
dill==0.4.0
einops==0.8.1
exceptiongroup==1.3.1
executing==2.2.1
fairscale==0.4.13
fastapi==0.122.0
ffmpy==1.0.0
filelock==3.20.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
gitdb==4.0.12
GitPython==3.1.45
gradio==6.0.1
gradio_client==2.0.0
groovy==0.1.2
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==8.7.0
iniconfig==2.3.0
ipdb==0.13.13
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
kiwisolver==1.4.9
lightning-utilities==0.15.2
lmdb==1.7.5
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
mdurl==0.1.2
modelscope==1.32.0
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.18
networkx==3.4.2
ninja==1.13.0
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
omegaconf==2.3.0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pillow==12.0.0
platformdirs==4.5.1
pluggy==1.6.0
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==3.20.3
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pydantic==2.12.4
pydantic_core==2.41.5
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pytest==7.2.0
pytest-split==0.8.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.6.0
pytz==2025.2
PyYAML==6.0.3
regex==2025.11.3
requests==2.32.5
rich==14.2.0
safehttpx==0.1.7
safetensors==0.7.0
scipy==1.15.3
semantic-version==2.10.0
sentencepiece==0.2.1
sentry-sdk==2.47.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.50.0
sympy==1.14.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
timm==1.0.22
tokenizers==0.22.1
tomli==2.3.0
tomlkit==0.13.3
torch==2.8.0
torchdiffeq==0.2.5
torchmetrics==1.8.2
torchvision==0.23.0
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
triton==3.4.0
typer==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wandb==0.23.1
wcwidth==0.2.14
webdataset==1.0.2
Werkzeug==3.1.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
